Refactor onboarding completion flow modules
